added code coverage logging. Further categorized the GamesController tests
