+++ tests/Feature/DeleteGamesTest.php
@@ -43,4 +43,32 @@ class DeleteGamesTest extends TestCase
         $this->delete($game->path())
             ->assertRedirect('/games');
     }
+
+    /** @test */
+    public function an_unauthenticated_user_may_not_delete_a_game()
+    {
+        $this->withExceptionHandling();
+        $game = create(Game::class);
+
+        $this->delete($game->path())
+             ->assertRedirect('/login');
+
+        $this->signIn();
+        $this->delete($game->path())
+             ->assertStatus(403);
+    }
+
+    /** @test */
+    public function an_authenticated_user_can_delete_a_game()
+    {
+        $this->signIn();
+        $game = create(Game::class, [
+            'user_id' => $this->user->id
+        ]);
+
+        $this->delete($game->path());
+
+        $this->assertDatabaseMissing('games', $game->toArray());
+    }
+
 }
